feat: update documentation for version 2.5.0 of the Next.js SDK
The upcoming 2.5.0 release of the Next.js SDK makes middleware significantly more important - it's now essential for efficient token refreshes.

To signify this, the Middleware section has been moved up from the bottom of the page towards the getting started/configuration section and wording has been change to denote that it's now a requirement, not optional.

The middleware matchers have been updated to the new 2.5.0 format as well.